# api-backend/fripay/public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FriPay API Documentation",
        "version": "1.0.0",
        "description": "API FriPay - Transferts inter-op\u00e9rateurs Mobile Money pour le B\u00e9nin. Endpoint de documentation: `\/api\/docs\/api`"
    },
    "servers": [
        {
            "url": "http:\/\/payments.fripay.local\/api"
        }
    ],
    "paths": {
...
        "\/v1\/qr\/generate": {
            "post": {
                "operationId": "offlineQr.generate",
                "description": "Cr\u00e9e un QR Code contenant une valeur mon\u00e9taire sign\u00e9e cryptographiquement\navec une cl\u00e9 Ed25519. Le QR Code peut \u00eatre partag\u00e9 hors-ligne (affichage\n\u00e0 l'\u00e9cran, capture d'\u00e9cran, etc.) et v\u00e9rifi\u00e9 sans connexion Internet.\n\nLe QR Code contient : UUID, montant, devise, cl\u00e9 publique de l'exp\u00e9diteur,\nhorodatage, et date d'expiration. Le tout est sign\u00e9 avec la cl\u00e9 secr\u00e8te\nEd25519 de l'exp\u00e9diteur.",
                "summary": "G\u00e9n\u00e9rer un QR Code sign\u00e9",
                "tags": [
                    "QR Codes Hors-ligne",
                    "OfflineQr"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "minimum": 100,
                                        "maximum": 500000
                                    },
                                    "currency": {
                                        "type": "string",
                                        "maxLength": 3
                                    },
                                    "expires_minutes": {
                                        "type": "integer",
                                        "minimum": 5,
                                        "maximum": 1440
                                    },
                                    "recipient_hint": {
                                        "type": [
                                            "string",
                                            "null"
                                        ],
                                        "maxLength": 100
                                    }
                                },
                                "required": [
                                    "amount"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "qr_code": {
                                            "type": "string"
                                        },
                                        "uuid": {
                                            "type": "string"
                                        },
                                        "amount": {
                                            "type": "string"
                                        },
                                        "currency": {
                                            "anyOf": [
                                                {
                                                    "type": "string"
                                                },
                                                {
                                                    "type": "string",
                                                    "enum": [
                                                        "XOF"
                                                    ]
                                                }
                                            ]
                                        },
                                        "expires_at": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string",
                                            "const": "active"
                                        }
                                    },
                                    "required": [
                                        "qr_code",
                                        "uuid",
                                        "amount",
                                        "currency",
                                        "expires_at",
                                        "status"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#\/components\/responses\/ValidationException"
                    },
                    "401": {
                        "$ref": "#\/components\/responses\/AuthenticationException"
                    }
                }
            }
        },
...
    },
...
}